Fix jittery status text
I've fixed the jittery issue with the status text field by removing the intermediary "Attempting to reconnect..." message that was causing the problem.

Here's what was happening:

1. The status text would show "Not connected. Retrying in X seconds"
2. When the timer finished, it would briefly show "Attempting to reconnect..."
3. Then very quickly change to either "Connected..." or back to "Not connected. Retrying..."

This rapid succession of text changes with different lengths was causing the abrupt height changes and jitter.

The solution:

1. I removed the CSS transitions we added earlier as they weren't fully solving the issue
2. I removed the code in `connection.js` that was setting the brief "Attempting to reconnect..." message

Now the status will transition directly from the countdown message to either the connected state or the next countdown message, eliminating the jittery effect completely. The intermediate status message wasn't providing essential information to the user anyway, so this is a cleaner solution.